% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFunctions.R
\name{read_tcx_directory}
\alias{read_tcx_directory}
\title{Read a whole directory of '.tcx' files.}
\usage{
read_tcx_directory(home_path)
}
\arguments{
\item{home_path}{The path to read the files from.}
}
\value{
Tidy data frame of all runs
}
\description{
Read a whole directory of '.tcx' files.
}
\seealso{
You can do a bulk download of all your Garmin runs at
\url{http://www.sideburn.org/garmin/}
}

