% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFunctions.R
\name{read_tcx_run}
\alias{read_tcx_run}
\title{Read a '.tcx' file.}
\usage{
read_tcx_run(path)
}
\arguments{
\item{path}{The path to read the file from}
}
\value{
Tidy data frame of the run
}
\description{
The function \code{read_tcx_run} reads a '.tcx' file from a given path and
returns a tidy data frame of the run.
}
\details{
In the \code{path}, you have to replace the '\' by '/',
 if you are using a windows system.
}

